requirements.txt
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.0
cycler==0.11.0
db-dtypes==1.1.1
Fiona==1.9.4.post1
fonttools==4.40.0
geopandas==0.13.2
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-bigquery==3.11.1
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
MetPy==1.5.0
ncBuilder==0.1.0
netCDF4==1.6.4
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
Pint==0.21.1
pkg_resources==0.0.0
platformdirs==3.6.0
pooch==1.7.0
proto-plus==1.22.2
protobuf==4.23.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
threadpoolctl==3.1.0
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.16
xarray==2023.1.0
zipp==3.15.0
PyKrige==1.7.2
global-land-mask==1.0.0
